+++ splitgraph/commands/__init__.py
@@ -5,7 +5,5 @@ Splitgraph public command API
 from splitgraph.commands._drawing import render_tree
 from splitgraph.commands.misc import get_log, cleanup_objects, init, rm
 from splitgraph.commands.mounting import mount
-from splitgraph.commands.provenance import provenance, image_hash_to_splitfile
 from splitgraph.commands.publish import publish
 from splitgraph.commands.push_pull import push, pull, clone
-from splitgraph.commands.tagging import get_current_head, get_all_hashes_tags, set_tag, resolve_image, delete_tag
