+++ splitgraph/commands/publish.py
@@ -8,9 +8,7 @@ from datetime import datetime
 from psycopg2.sql import SQL, Identifier
 
 from splitgraph._data.registry import publish_tag
-from splitgraph.commands.provenance import provenance
 from splitgraph.commands.push_pull import merge_push_params
-from splitgraph.commands.tagging import get_tagged_id
 from splitgraph.engine import get_engine, switch_engine
 
 PREVIEW_SIZE = 100
@@ -30,12 +28,13 @@ def publish(repository, tag, remote_engine_name=None, remote_repository=None, re
     :param include_table_previews: Whether to include data previews for every table in the image.
     """
     remote_engine_name, remote_repository = merge_push_params(repository, remote_engine_name, remote_repository)
-    image_hash = get_tagged_id(repository, tag)
+    image_hash = repository.get_tagged_id(tag)
     logging.info("Publishing %s:%s (%s)", repository, image_hash, tag)
 
-    dependencies = [((r.namespace, r.repository), i) for r, i in provenance(repository, image_hash)] \
+    image = repository.get_image(image_hash)
+    dependencies = [((r.namespace, r.repository), i) for r, i in image.provenance()] \
         if include_provenance else None
-    previews, schemata = _prepare_extra_data(image_hash, repository, include_table_previews)
+    previews, schemata = _prepare_extra_data(image, repository, include_table_previews)
 
     remote_engine = get_engine(remote_engine_name)
     try:
@@ -47,19 +46,18 @@ def publish(repository, tag, remote_engine_name=None, remote_repository=None, re
         remote_engine.close()
 
 
-def _prepare_extra_data(image_hash, repository, include_table_previews):
+def _prepare_extra_data(image, repository, include_table_previews):
     schemata = {}
     previews = {}
-    image_object = repository.get_image(image_hash)
-    for table_name in image_object.get_tables():
+    for table_name in image.get_tables():
         if include_table_previews:
             logging.info("Generating preview for %s...", table_name)
-            with repository.materialized_table(table_name, image_hash) as (tmp_schema, tmp_table):
+            with repository.materialized_table(table_name, image.image_hash) as (tmp_schema, tmp_table):
                 engine = get_engine()
                 schema = engine.get_full_table_schema(tmp_schema, tmp_table)
                 previews[table_name] = engine.run_sql(SQL("SELECT * FROM {}.{} LIMIT %s").format(
                         Identifier(tmp_schema), Identifier(tmp_table)), (PREVIEW_SIZE,))
         else:
-            schema = image_object.get_table_schema(table_name)
+            schema = image.get_table_schema(table_name)
         schemata[table_name] = [(cn, ct, pk) for _, cn, ct, pk in schema]
     return previews, schemata
