--- splitgraph/commands/tagging.py
@@ -1,140 +1,0 @@
-"""
-API functions for tagging images/getting tag information
-"""
-
-from psycopg2.sql import SQL, Identifier
-
-from splitgraph.config import SPLITGRAPH_META_SCHEMA
-from splitgraph.engine import ResultShape, get_engine
-from splitgraph.exceptions import SplitGraphException
-from .repository import repository_exists
-from .._data.common import ensure_metadata_schema, select, insert
-
-
-def get_current_head(repository, raise_on_none=True):
-    """
-    Gets the currently checked out image hash for a given repository.
-
-    :param repository: Repository
-    :param raise_on_none: Whether to raise an error or return None if the repository isn't checked out.
-    """
-    return get_tagged_id(repository, 'HEAD', raise_on_none)
-
-
-def get_tagged_id(repository, tag, raise_on_none=True):
-    """
-    Returns the image hash with a given tag in a given repository.
-
-    :param repository: Repository
-    :param tag: Tag. 'latest' is a special case: it returns the most recent image in the repository.
-    :param raise_on_none: Whether to raise an error or return None if the repository isn't checked out.
-    """
-    ensure_metadata_schema()
-    if not repository_exists(repository) and raise_on_none:
-        raise SplitGraphException("%s does not exist!" % str(repository))
-    engine = get_engine()
-
-    if tag == 'latest':
-        # Special case, return the latest commit from the repository.
-        result = engine.run_sql(select("images", "image_hash", "namespace = %s AND repository = %s")
-                                + SQL(" ORDER BY created DESC LIMIT 1"), (repository.namespace, repository.repository,),
-                                return_shape=ResultShape.ONE_ONE)
-        if result is None:
-            raise SplitGraphException("No commits found in %s!")
-        return result
-
-    result = engine.run_sql(select("tags", "image_hash", "namespace = %s AND repository = %s AND tag = %s"),
-                            (repository.namespace, repository.repository, tag),
-                            return_shape=ResultShape.ONE_ONE)
-    if result is None:
-        if raise_on_none:
-            schema = repository.to_schema()
-            if tag == 'HEAD':
-                raise SplitGraphException("No current checked out revision found for %s. Check one out with \"sgr "
-                                          "checkout %s image_hash\"." % (schema, schema))
-            else:
-                raise SplitGraphException("Tag %s not found in repository %s" % (tag, schema))
-        else:
-            return None
-    return result
-
-
-def get_all_hashes_tags(repository):
-    """
-    Gets all tagged images and their hashes in a given repository.
-
-    :param repository: Repository
-    :return: List of (image_hash, tag)
-    """
-    return get_engine().run_sql(select("tags", "image_hash, tag", "namespace = %s AND repository = %s"),
-                                (repository.namespace, repository.repository,))
-
-
-def set_tags(repository, tags, force=False):
-    """
-    Sets tags for multiple images.
-
-    :param repository: Repository
-    :param tags: List of (image_hash, tag)
-    :param force: Whether to remove the old tag if an image with this tag already exists.
-    """
-    for tag, image_id in tags.items():
-        if tag != 'HEAD':
-            set_tag(repository, image_id, tag, force)
-
-
-def set_tag(repository, image, tag, force=False):
-    """
-    Tags a given image. All tags are unique inside of a repository.
-
-    :param repository: Repository that the image belongs to
-    :param image: Image hash
-    :param tag: Tag to set. 'latest' and 'HEAD' are reserved tags.
-    :param force: Whether to remove the old tag if an image with this tag already exists.
-    """
-    engine = get_engine()
-    if engine.run_sql(select("tags", "1", "namespace = %s AND repository = %s AND tag = %s"),
-                      (repository.namespace, repository.repository, tag), return_shape=ResultShape.ONE_ONE) is None:
-        engine.run_sql(insert("tags", ("image_hash", "namespace", "repository", "tag")),
-                       (image, repository.namespace, repository.repository, tag),
-                       return_shape=None)
-    else:
-        if force:
-            engine.run_sql(SQL("UPDATE {}.tags SET image_hash = %s "
-                               "WHERE namespace = %s AND repository = %s AND tag = %s")
-                           .format(Identifier(SPLITGRAPH_META_SCHEMA)),
-                           (image, repository.namespace, repository.repository, tag),
-                           return_shape=None)
-        else:
-            raise SplitGraphException("Tag %s already exists in %s!" % (tag, repository.to_schema()))
-
-
-def delete_tag(repository, tag):
-    """
-    Deletes a tag from an image.
-
-    :param repository: Repository the image belongs to.
-    :param tag: Tag to delete.
-    """
-
-    # Does checks to make sure the tag actually exists, will raise otherwise
-    get_tagged_id(repository, tag)
-
-    get_engine().run_sql(SQL("DELETE FROM {}.tags WHERE namespace = %s AND repository = %s AND tag = %s")
-                         .format(Identifier(SPLITGRAPH_META_SCHEMA)),
-                         (repository.namespace, repository.repository, tag),
-                         return_shape=None)
-
-
-def resolve_image(repository, tag_or_hash, raise_on_none=True):
-    """Converts a tag or shortened hash to a full image hash that exists in the repository."""
-    try:
-        return repository.get_canonical_image_id(tag_or_hash)
-    except SplitGraphException:
-        try:
-            return get_tagged_id(repository, tag_or_hash)
-        except SplitGraphException:
-            if raise_on_none:
-                raise SplitGraphException("%s does not refer to either an image commit hash or a tag!" % tag_or_hash)
-            else:
-                return None
